verify numa

+++ src/cpp/include/list_scanning.h
@@ -538,9 +538,9 @@ inline void l2_blas(
             faiss::fvec_norms_L2sqr(norms_y, y + j0 * d, d, db_chunk);
 
             // use torch matmul
-            Tensor x_tensor = torch::from_blob((void*) x + i0 * d, {q_chunk, d}, torch::kFloat32);
-            Tensor y_tensor = torch::from_blob((void*) y + j0 * d, {db_chunk, d}, torch::kFloat32);
-            Tensor ip_tensor = torch::from_blob(ip_block, {q_chunk, db_chunk}, torch::kFloat32);
+            Tensor x_tensor = torch::from_blob((void*) x + i0 * d, {(int64_t) q_chunk, (int64_t) d}, torch::kFloat32);
+            Tensor y_tensor = torch::from_blob((void*) y + j0 * d, {(int64_t) db_chunk, (int64_t) d}, torch::kFloat32);
+            Tensor ip_tensor = torch::from_blob(ip_block, {(int64_t) q_chunk, (int64_t) db_chunk}, torch::kFloat32);
             torch::matmul_out(ip_tensor, x_tensor, y_tensor.transpose(0, 1));
 
             /* SGEMM */
